pyproject.toml
[build-system]
requires = ["hatchling>=1.26.0"]
build-backend = "hatchling.build"

[project]
name = "gemma4-capability-map"
version = "0.1.0"
description = "A white-box benchmark for reasoning, tool use, retrieval, and efficiency in Gemma-native agent stacks."
readme = "README.md"
requires-python = ">=3.13"
license = { text = "MIT" }
authors = [
  { name = "Codex" }
]
dependencies = [
  "pydantic>=2.10.0",
  "PyYAML>=6.0.2",
  "streamlit>=1.44.0",
]

[project.optional-dependencies]
hf = [
  "accelerate>=1.7.0",
  "torch>=2.7.0",
  "transformers>=4.52.0",
  "torchvision>=0.22.0",
  "sentence-transformers>=4.0.0",
]
mlx = [
  "mlx-lm>=0.30.0",
  "mlx-vlm>=0.3.9",
]
dev = [
  "pytest>=8.3.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/gemma4_capability_map"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
